put top-level tests into separate file
